Finished all precompilation and ability to toggle on and off, but haven't tested AOT blending and smoothing yet

+++ src/api/requests.py
@@ -21,7 +21,7 @@ class StillRequest(BaseModel):
 
 
 class AnimationRequest(StillRequest):
-    delta: int
+    delta: float
     frames: int
     fps: int
